pyproject.toml
[project]
name = "pythonproject-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "flask-cors>=5.0.1",
    "music21>=9.5.0",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
]
